Convery survey to .vue component
